# sqlbase: remove ColumnType from EncDatum