--- a/pkg/sql/distsqlrun/expr.go
+++ b/pkg/sql/distsqlrun/expr.go
@@ -109,7 +109,7 @@ func (eh *exprHelper) IndexedVarResolvedType(idx int) parser.Type {
 
 // IndexedVarEval is part of the parser.IndexedVarContainer interface.
 func (eh *exprHelper) IndexedVarEval(idx int, ctx *parser.EvalContext) (parser.Datum, error) {
-	err := eh.row[idx].EnsureDecoded(&eh.datumAlloc)
+	err := eh.row[idx].EnsureDecoded(&eh.types[idx], &eh.datumAlloc)
 	if err != nil {
 		return nil, err
 	}