@asubiotto 2017-10-16
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Reviewed 11 of 11 files at r1, 49 of 54 files at r2, 2 of 2 files at r3.
Review status: 50 of 55 files reviewed at latest revision, 4 unresolved discussions, all commit checks successful.

---

*[pkg/sql/distsqlrun/algebraic_set_op_test.go, line 83 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246#-Kw_nAQ3aglcoZeItvss:-Kw_nAQ3aglcoZeItvst:bv1kmat) ([raw file](https://github.com/cockroachdb/cockroach/blob/a38d30bcd1c31ff99bd3d0ca71033d249923ced4/pkg/sql/distsqlrun/algebraic_set_op_test.go#L83)):*
> ```Go
> 
> func runProcessors(tc testCase) (sqlbase.EncDatumRows, error) {
> 	types := []sqlbase.ColumnType{intType, intType}
> ```

Could use `twoIntCols` here.

---

*[pkg/sql/distsqlrun/mergejoiner_test.go, line 42 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246#-Kwa0eb626tvMuA9NLWp:-Kwa0eb626tvMuA9NLWq:b-ayegej) ([raw file](https://github.com/cockroachdb/cockroach/blob/b68ed89b085c340f095e1bd399c5b152f2170c6c/pkg/sql/distsqlrun/mergejoiner_test.go#L42)):*
> ```Go
> 		spec       MergeJoinerSpec
> 		outCols    []uint32
> 		leftTypes  []sqlbase.ColumnType
> ```

Consider using this method of creating types in `sorter_test` as well. It would be nice not to have different tests use different ways to create types. However, I also don't like that you would have to create the types manually and `sorter_test` has longer int rows than the global test variables you've introduced. Maybe you could create some function that takes a length and returns an intType slice of that length.

---

*[pkg/sql/sqlbase/table.go, line 428 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246#-Kw_zgvS3RZ8hNI85RjD:-Kw_zgvS3RZ8hNI85RjE:b-zapi5j) ([raw file](https://github.com/cockroachdb/cockroach/blob/b68ed89b085c340f095e1bd399c5b152f2170c6c/pkg/sql/sqlbase/table.go#L428)):*
> ```Go
> 	if len(values) != len(types) {
> 		panic(fmt.Sprintf("%d values, %d types", len(values), len(types)))
> 		//return nil, errors.Errorf("%d values, %d types", len(values), len(types))
> ```

Delete this line?

---

*[pkg/util/encoding/encoding.go, line 567 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246#-Kwa-6LTgtgp-PcT76Db:-Kwa-6LTgtgp-PcT76Dc:b-7ropln) ([raw file](https://github.com/cockroachdb/cockroach/blob/b68ed89b085c340f095e1bd399c5b152f2170c6c/pkg/util/encoding/encoding.go#L567)):*
> ```Go
> 		if i == -1 {
> 			panic(fmt.Sprintf("did not find terminator %#x in buffer %#x", e.escape, b))
> 			//return 0, errors.Errorf("did not find terminator %#x in buffer %#x", e.escape, b)
> ```

ditto

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246#-:-Kwa214iqjy5CtsTRbVA:bnfp4nl)*
<!-- Sent from Reviewable.io -->