@RaduBerinde 2017-10-16
Updated. Also removed the unnecessary `alloc` param from `DecodeIndexKey`(CC @richardwu)

---

Review status: 44 of 54 files reviewed at latest revision, 4 unresolved discussions, some commit checks failed.

---

*[pkg/sql/distsqlrun/algebraic_set_op_test.go, line 83 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246#-Kw_nAQ3aglcoZeItvss:-KwatzPYp6M3FHOoIGM2:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/a38d30bcd1c31ff99bd3d0ca71033d249923ced4/pkg/sql/distsqlrun/algebraic_set_op_test.go#L83)):*
<details><summary><i>Previously, asubiotto (Alfonso Subiotto Marqués) wrote…</i></summary><blockquote>

Could use `twoIntCols` here.
</blockquote></details>

Done.

---

*[pkg/sql/distsqlrun/mergejoiner_test.go, line 42 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246#-Kwa0eb626tvMuA9NLWp:-KwaxluBOXDDBcQZs6NA:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/b68ed89b085c340f095e1bd399c5b152f2170c6c/pkg/sql/distsqlrun/mergejoiner_test.go#L42)):*
<details><summary><i>Previously, asubiotto (Alfonso Subiotto Marqués) wrote…</i></summary><blockquote>

Consider using this method of creating types in `sorter_test` as well. It would be nice not to have different tests use different ways to create types. However, I also don't like that you would have to create the types manually and `sorter_test` has longer int rows than the global test variables you've introduced. Maybe you could create some function that takes a length and returns an intType slice of that length.
</blockquote></details>

Done.

---

*[pkg/sql/sqlbase/table.go, line 428 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246#-Kw_zgvS3RZ8hNI85RjD:-Kwatzv7AEoFlzN12R0W:beit15y) ([raw file](https://github.com/cockroachdb/cockroach/blob/b68ed89b085c340f095e1bd399c5b152f2170c6c/pkg/sql/sqlbase/table.go#L428)):*
<details><summary><i>Previously, asubiotto (Alfonso Subiotto Marqués) wrote…</i></summary><blockquote>

Delete this line?
</blockquote></details>

Oops, this panic was used to debug some issues, reverted.

---

*[pkg/util/encoding/encoding.go, line 567 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246#-Kwa-6LTgtgp-PcT76Db:-Kwau4kf6_6Zjx3wYAW1:bvth56w) ([raw file](https://github.com/cockroachdb/cockroach/blob/b68ed89b085c340f095e1bd399c5b152f2170c6c/pkg/util/encoding/encoding.go#L567)):*
<details><summary><i>Previously, asubiotto (Alfonso Subiotto Marqués) wrote…</i></summary><blockquote>

ditto
</blockquote></details>

reverted

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246#-:-KwayNRvOmtozx0ZvoN5:b-eal0tp)*
<!-- Sent from Reviewable.io -->