@asubiotto 2017-10-17
Review status: 1 of 54 files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[pkg/sql/distsqlrun/algebraic_set_op_test.go, line 83 at r5](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246#-KwepJrfjV3BiFiyZ293:-KwepJrfjV3BiFiyZ294:bd334g2) ([raw file](https://github.com/cockroachdb/cockroach/blob/28fecfdc7b42f99ea7fd8fc6a635c0db71fa11c4/pkg/sql/distsqlrun/algebraic_set_op_test.go#L83)):*
> ```Go
> 
> func runProcessors(tc testCase) (sqlbase.EncDatumRows, error) {
> 	types := twoIntCols
> ```

`twoIntCols` could be passed directly to `NewRowBuffer` here.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246)*
<!-- Sent from Reviewable.io -->