@RaduBerinde 2017-10-17
*[pkg/sql/distsqlrun/algebraic_set_op_test.go, line 83 at r5](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246#-KwepJrfjV3BiFiyZ293:-Kwevt-6uBj6UYF4u3iQ:bfssli) ([raw file](https://github.com/cockroachdb/cockroach/blob/28fecfdc7b42f99ea7fd8fc6a635c0db71fa11c4/pkg/sql/distsqlrun/algebraic_set_op_test.go#L83)):*
<details><summary><i>Previously, asubiotto (Alfonso Subiotto Marqués) wrote…</i></summary><blockquote>

`twoIntCols` could be passed directly to `NewRowBuffer` here.
</blockquote></details>

it's used below too

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19246)*
<!-- Sent from Reviewable.io -->